test: isolate the resolved-badge getNodeDefs case in its own file
The badge fetch is a session singleton settled by the first
getNodeDefs() call, so the resolved-badge assertion needed to run
first in the shared file. Give it its own module registry instead,
matching the pending-pricing counterpart, and drop the ordering
requirement.
